add typescript + prettier' gst
