+++ package.json
@@ -7,6 +7,10 @@
     "@fortawesome/free-solid-svg-icons": "^6.2.1",
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@material-tailwind/react": "^1.2.5",
+    "@types/jest": "^29.4.0",
+    "@types/node": "^18.13.0",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.10",
     "axios": "^1.2.1",
     "react": "^18.2.0",
     "react-animated-heart": "^0.0.7",
@@ -14,13 +18,15 @@
     "react-router-dom": "^6.4.4",
     "react-scripts": "^5.0.1",
     "react-to-print": "^2.14.10",
-    "twitter-like-button": "^0.1.6"
+    "twitter-like-button": "^0.1.6",
+    "typescript": "^4.9.5"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build && echo '/* /index.html 200' | cat >build/_redirects",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "check-types": "tsc --noEmit"
   },
   "eslintConfig": {
     "extends": [
@@ -44,6 +50,9 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^14.4.3",
+    "@tsconfig/create-react-app": "^1.0.3",
+    "prettier": "^2.8.4",
+    "prettier-plugin-tailwindcss": "^0.2.2",
     "tailwindcss": "^3.2.4"
   }
 }
