+++ src/App.jsx
@@ -1,34 +1,35 @@
-import "./App.css";
-import {Routes, Route} from "react-router-dom";
-import React, {useEffect, useState} from "react";
-import axios from "axios";
-import HomePage from "./pages/HomePage/HomePage";
-import ProfilePage from "./pages/ProfilePage/ProfilePage";
-import SignupPage from "./pages/SignupPage/SignupPage";
-import LoginPage from "./pages/LoginPage/LoginPage";
-import Footer from "./components/Navbar/Footer";
-import Navbar from "./components/Navbar/Navbar";
-import IsPrivate from "./components/IsPrivate/IsPrivate";
-import IsAnon from "./components/IsAnon/IsAnon";
-import Create from "./pages/CRUD/Create";
-import Details from "./pages/CRUD/Details";
-import SingleRecipe from "./pages/CRUD/SingleRecipe";
-import RecipesHome from "./pages/CRUD/RecipesHome";
+import './App.css'
+import { Routes, Route } from 'react-router-dom'
+import React, { useEffect, useState } from 'react'
+import axios from 'axios'
+import HomePage from './pages/HomePage/HomePage'
+import ProfilePage from './pages/ProfilePage/ProfilePage'
+import SignupPage from './pages/SignupPage/SignupPage'
+import LoginPage from './pages/LoginPage/LoginPage'
+import Footer from './components/Navbar/Footer'
+import Navbar from './components/Navbar/Navbar'
+import IsPrivate from './components/IsPrivate/IsPrivate'
+import IsAnon from './components/IsAnon/IsAnon'
+import Create from './pages/CRUD/Create'
+import Details from './pages/CRUD/Details'
+import SingleRecipe from './pages/CRUD/SingleRecipe'
+import RecipesHome from './pages/CRUD/RecipesHome'
+import NotFoundPage from './pages/NotFoundPage/NotFoundPage'
 //d
-const API_URL = process.env.REACT_APP_SERVER_URL || "http://localhost:5005";
+const API_URL = process.env.REACT_APP_SERVER_URL || 'http://localhost:5005'
 
 function App() {
-    const [data, setData] = useState([]);
+    const [data, setData] = useState([])
 
     const refresh = () => {
         axios
             .get(`${API_URL}/pages/CRUD/details`)
-            .then((response) => setData(response.data), console.log(data));
-    };
+            .then((response) => setData(response.data), console.log(data))
+    }
     useEffect(() => {
-        refresh();
-    }, []);
-    console.log(data);
+        refresh()
+    }, [])
+    console.log(data)
 
     return (
         <div className="App ">
@@ -89,11 +90,13 @@ function App() {
                     path="/edit/:id"
                     element={<Create refresh={refresh} data={data} />}
                 />
+
+                <Route path="*" element={<NotFoundPage />}></Route>
             </Routes>
 
             <Footer />
         </div>
-    );
+    )
 }
 
-export default App;
+export default App
