+++ tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@tsconfig/create-react-app/tsconfig.json"
+}
